1321 move routing logic to its own file so that it is a child component of the AppProvider component and will thus have access to the auth state
